Rework template variables
Each dashboard now gets their own metric and label for the template
variables. Having the values retrieved from a single metric, which would
only be shown on one dashboard, would probably not have caused any
issues. But it does feel weird to take values from a metric to filter a
dashboard that does not use this metric at all.